removed temp references
