+++ src/Core/Vec2.h
@@ -2,7 +2,9 @@
 #define VEC2_H_
 
 #include <cmath>
-const static float GRADUS = 3.14159265359f / 180;
+
+const static float PI_F = 3.14159265358979f;
+const static float GRADUS = PI_F / 180.0f;
 
 class Vec2 {
 public:
@@ -15,7 +17,7 @@ public:
     Vec2 GetNormalized() const;
     Vec2 GetRotated(const double& angle) const;
 
-    const float& Dot(const Vec2& right);
+    float Dot(const Vec2& right);
 
     Vec2& operator=(const Vec2& right);
     bool operator==(const Vec2& right) const;
@@ -28,5 +30,93 @@ public:
     friend Vec2& operator-=(Vec2& left, const Vec2& right);
     friend Vec2& operator*=(Vec2& left, const float& scale);
 };
+
+inline
+Vec2::Vec2() : x(0), y(0) {
+}
+
+inline
+Vec2::Vec2(const float& x, const float& y) : x(x), y(y) {
+}
+
+inline
+float Vec2::GetLength() const {
+    return sqrt(x * x + y * y);
+}
+
+inline
+Vec2 Vec2::GetNormalized() const {
+    float len = this->GetLength();
+    return Vec2(this->x / len, this->y / len);
+}
+
+inline
+Vec2 Vec2::GetRotated(const double& angle) const {
+    float rad = (angle) * GRADUS;
+    float sn = sin(rad);
+    float cs = cos(rad);
+    return Vec2((x * cs - y * sn), (x * sn + y * cs));
+}
+
+inline
+float Vec2::Dot(const Vec2& right){
+    return (this->x * right.x) + (this->y * right.y);
+}
+
+inline
+Vec2& Vec2::operator=(const Vec2& right) {
+    //проверка на самоприсваивание
+    if (this != &right) {
+        this->x = right.x;
+        this->y = right.y;
+    }
+    return *this;
+}
+
+inline
+bool Vec2::operator==(const Vec2& right) const {
+    return (this->x == right.x) && (this->y == right.y);
+}
+
+inline
+const Vec2 operator+(const Vec2& left, const Vec2& right) {
+    return Vec2(left.x + right.x, left.y + right.y);
+}
+
+inline
+const Vec2 operator-(const Vec2& left, const Vec2& right) {
+    return Vec2(left.x - right.x, left.y - right.y);
+}
+
+inline
+const Vec2 operator*(const Vec2& left, const float& scale) {
+    return Vec2(left.x * scale, left.y * scale);
+}
+
+inline
+float operator*(const Vec2& left, const Vec2& right) {
+    return left.x * right.x + left.y * right.y;
+}
+
+inline
+Vec2& operator+=(Vec2& left, const Vec2& right) {
+    left.x += right.x;
+    left.y += right.y;
+    return left;
+}
+
+inline
+Vec2& operator-=(Vec2& left, const Vec2& right) {
+    left.x -= right.x;
+    left.y -= right.y;
+    return left;
+}
+
+inline
+Vec2& operator*=(Vec2& left, const float& scale) {
+    left.x *= scale;
+    left.y *= scale;
+    return left;
+}
 
 #endif /* VEC2_H_ */
